requirements.txt
apache-airflow[amazon, azure]==2.10.3
azure-storage-blob==12.23.1
azure-identity==1.19.0
pandas==2.2.3
python-dotenv==1.0.1
scipy==1.14.1
tqdm==4.67.0